Updated the default min and max value for probes, added HVAC off mode and more actions and other changes. (#116)
* Update the Probe Set Temp Min and Max to Correct Default Values

The probe 1 and 2 targe temp min and max values are not correct. My PBV4PS2 min and max is 50 to 250. I set this as default because its a real world min and max for a probe temp and probably consistent between models.

* Added HVACMode.Off, Added cooling and off actions.

Allows the smoker to be set to off from the climate card. The heat mode button can be disabled in the climate so its not seen making it a nice option because pitboss dose not allow remote turn on.

Added an off and cooling action.

* Disable ON switch if module is not on.

* Added Literal Type

* Make requested changes.

* Update available for target probe.

* Update available property for on switch.

* Update available property for probes.

* Added available for both the power switch and the primer switch

* Added available for both the power switch and the primer switch

* Remove moduleIsOn Available check from probe targets.